more work done on the structuring of the tool. Prepping the next stage of dev...
---

Apply the template? Nah, that must have already happened in the render phase.  If we have special 'generated content only' pages, such as the index page to a catalog site, that *still* is done through a (possibly empty, except for some metadata perhaps) `index.html` or README.md or other *content* file: the custom code can decide which bit of the template collective to apply to each page, so that would then apply a 'overview/index/landing' page template to such a (hypothetical) page.

Hence we come to the conclusion now: every page being written is written by a (possibly empty) content page. If the content page is absent, the page simply is NOT generated.

Meanwhile, the custom code decides which template file((s) are applied to each content item in the allFiles list.

---

Yes, this means we'll have content pages for the 404 and other landing pages too. If they are absent, we do not have those landing pages. Simple as that.